+++ trunk/rRootage/src/foecommand.cc
@@ -33,35 +33,35 @@ FoeCommand::FoeCommand(BulletMLState *state, Foe *f)
 
 FoeCommand::~FoeCommand() {}
 
-double FoeCommand::getBulletDirection() {
-  return (double)foe->d*360/DIV;
+float FoeCommand::getBulletDirection() {
+  return (float)foe->d*360/DIV;
 }
 
-double FoeCommand::getAimDirection() {
+float FoeCommand::getAimDirection() {
   int d = getPlayerDeg(foe->pos.x, foe->pos.y);
   if ( foe->xReverse == -1 ) d = (-d)&1023;
-  return ((double)d*360/DIV);
+  return ((float)d*360/DIV);
 }
 
-double FoeCommand::getBulletSpeed() {
-  return ((double)foe->spd)/COMMAND_SCREEN_SPD_RATE;
+float FoeCommand::getBulletSpeed() {
+  return ((float)foe->spd)/COMMAND_SCREEN_SPD_RATE;
 }
 
-double FoeCommand::getDefaultSpeed() {
+float FoeCommand::getDefaultSpeed() {
   return 1;
 }
 
-double FoeCommand::getRank() {
+float FoeCommand::getRank() {
   return foe->rank;
 }
 
-void FoeCommand::createSimpleBullet(double direction, double speed) {
+void FoeCommand::createSimpleBullet(float direction, float speed) {
   int d = (int)(direction*DIV/360); d &= (DIV-1);
   addFoeNormalBullet(foe, d, (int)(speed*COMMAND_SCREEN_SPD_RATE), foe->color+1);
   foe->fireCnt++;
 }
 
-void FoeCommand::createBullet(BulletMLState* state, double direction, double speed) {
+void FoeCommand::createBullet(BulletMLState* state, float direction, float speed) {
   int d = (int)(direction*DIV/360); d &= (DIV-1);
   addFoeActiveBullet(foe, d, (int)(speed*COMMAND_SCREEN_SPD_RATE), foe->color+1, state);
   foe->fireCnt++;
@@ -75,26 +75,26 @@ void FoeCommand::doVanish() {
   removeFoeCommand(foe);
 }
 
-void FoeCommand::doChangeDirection(double d) {
+void FoeCommand::doChangeDirection(float d) {
   foe->d = (int)(d*DIV/360);
 }
 
-void FoeCommand::doChangeSpeed(double s) {
+void FoeCommand::doChangeSpeed(float s) {
   foe->spd = (int)(s*COMMAND_SCREEN_SPD_RATE);
 }
 
-void FoeCommand::doAccelX(double ax) {
+void FoeCommand::doAccelX(float ax) {
   foe->vel.x = (int)(ax*COMMAND_SCREEN_VEL_RATE);
 }
 
-void FoeCommand::doAccelY(double ay) {
+void FoeCommand::doAccelY(float ay) {
   foe->vel.y = (int)(ay*COMMAND_SCREEN_VEL_RATE);
 }
 
-double FoeCommand::getBulletSpeedX() {
-  return ((double)foe->vel.x/COMMAND_SCREEN_VEL_RATE);
+float FoeCommand::getBulletSpeedX() {
+  return ((float)foe->vel.x/COMMAND_SCREEN_VEL_RATE);
 }
 
-double FoeCommand::getBulletSpeedY() {
-  return ((double)foe->vel.y/COMMAND_SCREEN_VEL_RATE);
+float FoeCommand::getBulletSpeedY() {
+  return ((float)foe->vel.y/COMMAND_SCREEN_VEL_RATE);
 }
